requirements.txt
Django==5.0.6
djangorestframework==3.15.2
channels==4.1.0
# For production realtime scaling use Redis:
# channels-redis==4.2.0
